Initial workings of this application, to show the report descriptor
Obviously more is planned, but this is getting the basis down for a simple MVVM architecture and interface for showing all info about USB HIDs connected to the computer.

In the future, more tools based on general data will be available.

I'll figure out dependencies more solidly soon, and perhaps get rid of the xcworkspace.