Rough implementation of the Reports tab
I think this will change a lot in the future. There should be a split view, with the left being a nicer summary of each report, and clicking on one leads to the right side showing more information about the `HIDValue` and its children elements itself.

Though this was a nice introduction to NSOutlineView. It's a bit annoying to use with the odd choice of dataSource/delegate methods. However using the initial version of OutlineViewRow from https://stackoverflow.com/questions/45373039/how-to-program-a-nsoutlineview helped a lot. I didn't like connecting so many things inside of IB so I found a way to make it work nicely with the dataSource and the delegate. I think it's much cleaner (rather than a lot of linear if statements for structured data). And it's extendable for the future (dedicated closures to supply the cell view data, perhaps).